feat: support matching workspaces
Add new Type to the rule. Supported types are
"window" and "workspace".

Match the window as usual, by Title or AppID.
Match the workspace by Name or Output.

Handle the dynamic setting of action IDs a bit better.
Pass in the possible keys that an action might have, and set the
IDs accordingly, if the action contains the defined possible key,
use that as the ID.

Update and add tests.