feat: make scratchpad configurable (#11)
Add new configuration for the scratch show command, showScratchpadActions.
These actions are performed on the window that was shown from the scratchpad workspace.
This is in addition to the default actions of MoveWindowToWorkspace and FocusWindow.

Update README with necessary details of the configuration for the scratchpad show actions. 
Update example config.json.