feat: add support for listening on custom events
Events can now be specified in the config. The configured events
will be handled in the case when there are no specific events being
handled otherwise. The current implementation is very simple, i.e.
specify an event to listen to, and some actions to do when the event
happens.
Currently no conditional checks are performed on the event/action to
be run.
Update the EventStream return type from any to Event.