feat: add getter for possible keys for an event
The GetPossibleKeys() on an event returns the models.PossibleKeys{} with
the ID/WorkspaceID/ActiveWindowID populated, depending on whether or not
the field exists in the event.

We can now use the possible keys to determine the action keys more
reliably, since it's per event we get them. Also we do not have to use
reflection on the event's possible keys, since it's quite slow