feat: add action configuration for custom event actions
Add a new field, "when", to the custom events' actions. The field
is a condition in expr-lang, which evaluates the condition against the
event we're listening to.
E.g. the event WindowUrgencyChanged has an ID and an Urgent field, so
the "when"-condition can be "event.Urgent == true", or just
"event.Urgent" which evaluates to true, if the window's urgency was set.