feat: customizable events (#12)
* chore: move setuintfield and setstringfield to common

These are functions we may want to use elsewhere.

* feat: add getter for possible keys for an event

The GetPossibleKeys() on an event returns the models.PossibleKeys{} with
the ID/WorkspaceID/ActiveWindowID populated, depending on whether or not
the field exists in the event.

We can now use the possible keys to determine the action keys more
reliably, since it's per event we get them. Also we do not have to use
reflection on the event's possible keys, since it's quite slow

* feat: add support for listening on custom events

Events can now be specified in the config. The configured events
will be handled in the case when there are no specific events being
handled otherwise. The current implementation is very simple, i.e.
specify an event to listen to, and some actions to do when the event
happens.

Currently no conditional checks are performed on the event/action to
be run.

Update the EventStream return type from any to Event.

* fix: move parsing of the raw actions to a getter

Remove all actionList references, and use the ActionsFromRaw function
instead.

Add latest version changes to README.md and update the
examples/config.json to include the new fields.