feat: support conditional actions (#13)
* feat: add action configuration for custom event actions

Add a new field, "when", to the custom events' actions. The field
is a condition in expr-lang, which evaluates the condition against the
event we're listening to.
E.g. the event WindowUrgencyChanged has an ID and an Urgent field, so
the "when"-condition can be "event.Urgent == true", or just
"event.Urgent" which evaluates to true, if the window's urgency was set.

* feat: evaluate action condition and perform if it evaluates to true

When listening on the custom events, we evaluate the given condition for
the action, and if it evaluates to true, we perform the action.
If the expression evaluates to false, we don't do anything.

If there is no condition, we always perform the action.

* fix: don't set field if it already has a value

When dynamically setting the fields, first check if it already
has a value set, e.g. coming from the configuration, and if it does,
don't override it.

* docs(readme): update readme

Add the new conditional action feature documentation and
update the example configuration.

* feat: support conditional actions in rules as well

Add the conditional actions to the window/workspace matching
rules as well.

* docs: update readme and example config

Add the conditional actions examples for rule matching as well.