feat: show scratchpad windows in launcher (#20)
If there are multiple windows on the scratchpad workspace, open a launcher (fuzzel by default) and list the available windows so the user can choose the window to bring to the current workspace.

Add RunCommand function to enable shell commands to be run.
Add validation on the command to be run, to check for malicious commands.
Add tests for command validation.

Refactoring:
If a command has sub-commands, move the commands to a sub-folder for clearer separation.
Use jj instead of git to figure out the version and commit hash during just install.